+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This file is a history of the changes made to idearium-lib.
 ## Unreleased
 
 - Add `MandrillEmail` to common.
+- Replace `kue.process` with `kue.processJobs`. This now automates processing all jobs in the `jobs` directory.
 
 ## 1.0.0-alpha.14